Move Message type to net_relay module
Out of the net_relay::udp module, so it can be used by the TCP
implementation.
