+++ remote/src/net_relay/udp.rs
@@ -1,9 +1,8 @@
 //! Module with the UDP implementation of the net relay.
 
 use std::convert::TryFrom;
-use std::marker::PhantomData;
+use std::io;
 use std::net::SocketAddr;
-use std::{fmt, io};
 
 use heph::actor::messages::Terminate;
 use heph::actor::{self, NoMessages};
@@ -11,11 +10,11 @@ use heph::net::UdpSocket;
 use heph::rt::{self, Signal};
 use heph::util::either;
 use log::warn;
-use serde::de::{self, Deserialize, DeserializeOwned, Deserializer, MapAccess, Visitor};
-use serde::ser::{Serialize, SerializeStruct, Serializer};
+use serde::de::DeserializeOwned;
+use serde::ser::Serialize;
 
-use crate::net_relay::uuid::{Uuid, UuidGenerator};
-use crate::net_relay::{Route, Serde};
+use crate::net_relay::uuid::UuidGenerator;
+use crate::net_relay::{Message, Route, Serde};
 
 const MAX_PACKET_SIZE: usize = 1 << 16; // ~65kb.
 
@@ -116,117 +115,6 @@ where
             // Error receiving a packet.
             Err(Err(err)) => return Err(err),
         }
-    }
-}
-
-/// Message type that wraps the user's message `M`.
-struct Message<M> {
-    uuid: Uuid,
-    msg: M,
-}
-
-// NOTE: manually implementing this instead of deriving to not pull in a bunch
-// of dependencies.
-impl<'de, M> Deserialize<'de> for Message<M>
-where
-    M: Deserialize<'de>,
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        enum Field {
-            Uuid,
-            Msg,
-        }
-
-        impl<'de> Deserialize<'de> for Field {
-            fn deserialize<D>(deserializer: D) -> Result<Field, D::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                struct FieldVisitor;
-
-                impl<'de> Visitor<'de> for FieldVisitor {
-                    type Value = Field;
-
-                    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                        formatter.write_str("`uuid` or `message`")
-                    }
-
-                    fn visit_str<E>(self, value: &str) -> Result<Field, E>
-                    where
-                        E: de::Error,
-                    {
-                        match value {
-                            "uuid" => Ok(Field::Uuid),
-                            "message" => Ok(Field::Msg),
-                            _ => Err(de::Error::unknown_field(value, FIELDS)),
-                        }
-                    }
-                }
-
-                deserializer.deserialize_identifier(FieldVisitor)
-            }
-        }
-
-        struct MessageVisitor<M>(PhantomData<M>);
-
-        impl<'de, M> Visitor<'de> for MessageVisitor<M>
-        where
-            M: Deserialize<'de>,
-        {
-            type Value = Message<M>;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("struct Message")
-            }
-
-            fn visit_map<V>(self, mut map: V) -> Result<Message<M>, V::Error>
-            where
-                V: MapAccess<'de>,
-            {
-                let mut uuid = None;
-                let mut msg = None;
-                while let Some(key) = map.next_key()? {
-                    match key {
-                        Field::Uuid => {
-                            if uuid.is_some() {
-                                return Err(de::Error::duplicate_field("uuid"));
-                            }
-                            uuid = Some(map.next_value()?);
-                        }
-                        Field::Msg => {
-                            if msg.is_some() {
-                                return Err(de::Error::duplicate_field("message"));
-                            }
-                            msg = Some(map.next_value()?);
-                        }
-                    }
-                }
-                let uuid = uuid.ok_or_else(|| de::Error::missing_field("uuid"))?;
-                let msg = msg.ok_or_else(|| de::Error::missing_field("message"))?;
-                Ok(Message { uuid, msg })
-            }
-        }
-
-        const FIELDS: &'static [&'static str] = &["uuid", "message"];
-        deserializer.deserialize_struct("Message", FIELDS, MessageVisitor(PhantomData))
-    }
-}
-
-impl<M> Serialize for Message<M>
-where
-    M: Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut state = serializer.serialize_struct("Message", 2)?;
-        state.serialize_field("uuid", &self.uuid)?;
-        state.serialize_field("message", &self.msg)?;
-        state.end()
     }
 }
